@@ -4,8 +4,8 @@
   "private": true,
   "dependencies": {
     "bootstrap": "^5.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0-alpha-bdd6d5064-20211001",
+    "react-dom": "^18.0.0-alpha-bdd6d5064-20211001",
     "react-error-boundary": "^3.1.3",
     "react-router-dom": "^6.0.0-beta.5",
     "swr": "^1.0.1",
